+++ beobench/integration/rllib.py
@@ -1,19 +1,20 @@
 """RLlib integration in beobench."""
 
+import json
 import ray.tune
 import ray.tune.integration.wandb
 import ray.tune.integration.mlflow
+import wandb
+import uuid
+import os
 
 import beobench.utils
 import beobench.experiment.config_parser
+from beobench.constants import RAY_LOCAL_DIR_IN_CONTAINER, CONTAINER_DATA_DIR
 
 
 def run_in_tune(
     config: dict,
-    wandb_project: str = None,
-    wandb_entity: str = None,
-    mlflow_name: str = None,
-    use_gpu: bool = False,
 ) -> ray.tune.ExperimentAnalysis:
     """Run beobench experiment.
 
@@ -37,32 +38,47 @@ def run_in_tune(
     Returns:
         ray.tune.ExperimentAnalysis: analysis object from experiment.
     """
-    if wandb_project and wandb_entity:
-        callbacks = [_create_wandb_callback(wandb_project, wandb_entity)]
-    elif wandb_project or wandb_entity:
+    run_id = uuid.uuid4().hex
+
+    if config["general"]["wandb_project"] and config["general"]["wandb_entity"]:
+
+        callbacks = [
+            ray.tune.integration.wandb.WandbLoggerCallback(
+                project=config["general"]["wandb_project"],
+                entity=config["general"]["wandb_entity"],
+                group=config["general"]["wandb_group"],
+                id=run_id,
+            )
+        ]
+    elif config["general"]["wandb_project"] or config["general"]["wandb_entity"]:
         raise ValueError(
             "Only one of wandb_project or wandb_entity given, but both required."
         )
-    elif mlflow_name:
-        callbacks = [_create_mlflow_callback(mlflow_name)]
+    elif config["general"]["mlflow_name"]:
+        callbacks = [_create_mlflow_callback(config["general"]["mlflow_name"])]
     else:
         callbacks = []
+
+    if config["general"]["log_episode_data_from_rllib"]:
+        output_dir = (CONTAINER_DATA_DIR / "outputs" / f"outputs-{run_id}").absolute()
+        config["agent"]["config"]["config"]["output"] = str(output_dir)
 
     # combine the three incomplete ray tune experiment
     # definitions into a single complete one.
     rllib_config = beobench.experiment.config_parser.create_rllib_config(config)
 
     # change RLlib setup if GPU used
-    if use_gpu:
+    if config["general"]["use_gpu"]:
         rllib_config["config"]["num_gpus"] = 1
 
     # register the problem environment with ray tune
-    # env_creator is a module available in experiment containers
-    import env_creator  # pylint: disable=import-outside-toplevel,import-error
+    # provider is a module available in experiment containers
+    # pylint: disable=import-outside-toplevel,import-error
+    from beobench.experiment.provider import create_env
 
     ray.tune.registry.register_env(
         rllib_config["config"]["env"],
-        env_creator.create_env,
+        create_env,
     )
 
     # if run in notebook, change the output reported throughout experiment.
@@ -75,29 +91,28 @@ def run_in_tune(
     analysis = ray.tune.run(
         progress_reporter=reporter,
         callbacks=callbacks,
+        local_dir=RAY_LOCAL_DIR_IN_CONTAINER,
         **rllib_config,
     )
 
+    if (
+        config["general"]["wandb_project"]
+        and config["general"]["log_episode_data_from_rllib"]
+    ):
+
+        wandb.init(
+            id=run_id,
+            project=config["general"]["wandb_project"],
+            entity=config["general"]["wandb_entity"],
+            group=config["general"]["wandb_group"],
+        )
+        print("Beobench: wandb run id", wandb.run.id)
+
+        output_file = output_dir / os.listdir(output_dir)[0]
+        data = get_cross_episodes_data(output_file)
+        log_eps_data_to_wandb(data, wandb_run_id=wandb.run.id)
+
     return analysis
-
-
-def _create_wandb_callback(
-    wandb_project: str,
-    wandb_entity: str,
-):
-    """Create an RLlib weights and biases (wandb) callback.
-
-    Args:
-        wandb_project (str): name of wandb project.
-        wandb_entity (str): name of wandb entity that owns project.
-
-    Returns:
-        : a wandb callback
-    """
-    wandb_callback = ray.tune.integration.wandb.WandbLoggerCallback(
-        project=wandb_project, log_config=True, entity=wandb_entity
-    )
-    return wandb_callback
 
 
 def _create_mlflow_callback(
@@ -115,3 +130,62 @@ def _create_mlflow_callback(
         experiment_name=mlflow_name, tracking_uri="file:/root/ray_results/mlflow"
     )
     return mlflow_callback
+
+
+def get_cross_episodes_data(path: str) -> dict:
+    """Get concatenated episode data from RLlib output.
+
+    This currently only concatenates data from info variable.
+    It further assumes that each info returned by step() in the env is
+    a dict of dicts.
+
+    Args:
+        path (str): path to RLlib output json file.
+
+    Returns:
+        dict: dictionary with episode data
+    """
+    # TODO: add non-info data to this logging procedure
+
+    # Open RLlib output
+    outputs = []
+    with open(path, encoding="UTF-8") as json_file:
+        for line in json_file.readlines():
+            outputs.append(json.loads(line))
+
+    # Get all keys of observations saved in info dict
+    all_obs_keys = []
+    for info_key in outputs[0]["infos"][0].keys():
+        all_obs_keys += list(outputs[0]["infos"][0][info_key].keys())
+
+    # Create empty (flat) dict of obs saved in info dict
+    eps_dict = {obs_key: [] for obs_key in all_obs_keys}
+
+    # Add data to dict, one step at a time
+    for output in outputs[:]:
+        for info in output["infos"]:
+            for info_key in info.keys():
+                obs_keys = outputs[0]["infos"][0][info_key].keys()
+                for obs_key in obs_keys:
+                    eps_dict[obs_key].append(info[info_key][obs_key])
+
+    return eps_dict
+
+
+def log_eps_data_to_wandb(eps_dict: dict, wandb_run_id: str) -> None:
+    """Log episode data to wandb.
+
+    To be used with concatenated episode data from get_cross_episodes_data().
+
+    Args:
+        eps_dict (dict): episode data
+        wandb_run_id (str): unique wandb run id to attach the data to.
+
+    """
+
+    wandb.init(id=wandb_run_id)
+
+    eps_dict_len = len(list(eps_dict.values())[0])
+    for i in range(eps_dict_len):
+        single_eps_dict = {key: values[i] for key, values in eps_dict.items()}
+        wandb.log(single_eps_dict)